// src/interfaces/client.ts
export interface Client {
  _idVendedor: string
  Apellidos: string
  ApellidosVendedor: string
  Cedula: string
  Codigo: number
  CodigoPostal: string
  CodigoZona: string
  Contacto1: string
  Contacto2: string
  Direccion: string
  email: string
  Empresa: string
  Estado: string
  Nombres: string
  NombresVendedor: string
  NumeroAdministrativo: string
  RIF: string
  SaldoFavor: number
  TipoPrecio: string
  Zona: string
}